team_2_music_back/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
redis
boto3
python-jose[cryptography]
python-dotenv
passlib[bcrypt]
httpx
pydantic-settings
python-multipart
pytest
pytest-asyncio
